backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
flask-jwt-extended
flask-mail
python-dotenv
flask-cors
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
selenium==4.15.2
fake-useragent==1.4.0
celery==5.3.4
redis==5.0.1
undetected-chromedriver==3.5.4